backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
alembic==1.18.4
python-jose[cryptography]==3.5.0
passlib[argon2]==1.7.4
email-validator==2.3.0
httpx==0.28.1
